requirements.txt
streamlit==1.38.0
langchain==0.3.1
langchain-core==0.3.6
langchain-openai==0.2.1
langchain-community==0.3.1
python-dotenv==1.0.1
